requirements.txt
# These are requirements for working on this package in isolation.
# Package dependencies are managed in the pyproject.toml
Django>=6.0
pre_commit>=4.5
pytest-cov>=7.0.0
pytest-django>=4.11.1
